Eeppl: grammar now generated with make_fun.py
which was extended to add a namespace prefix
in front of function names. So we get lin-statements like:

    Eeppl__Adila_Tagakuumlla = "Adila-Tagaküla";

Prefixes are needed to avoid possible clashes when
grammars are inherited, e.g. "Roo" is both a streetname
and a village name.
Using plain "Roo" as a function name would cause a clash
in an address grammar that inherits both street and village names.