Added: a very general unit conversion template
which supports sentences like:

    kaks eurot veerand tunnis teeb pool päevas veerand dollarites

Use case: you're in a country that uses a foreign currency and you want to
pay for parking. The cost of parking is reported in the foreign
currency for some duration (e.g. 2 EUR per 15 min).
You want to park for a different duration (half a day)
and you want to know the price in your own currency (USD). Furthermore, you have
only certain bills/coins (e.g. quarters), and (as the final result) you want
to know how many of these bills/coins you need to exchange.

The App-concrete grammar generates an expression that mixes unit conversion
with standard multiplication/division. Google can sometimes evaluate such
expressions (but not always), e.g. this works in Google:

    (EUR in USD) / (km in m)

This is a very preliminary implementation, things to be done:

  * generalize this template to all possible type pairs
  * correct morph. forms for numerals
  * allow for simpler versions where some of the four involved numbers are equal to one, and some of the units (e.g. currency) match, e.g. "kaks eurot veerand tunnis päevas")